move PushManifestList and all its utils into pkg
